FORGE-1525: Move EntityIdType
