+++ javaee/api/src/main/java/org/jboss/forge/addon/javaee/jpa/EntityIdType.java
@@ -4,7 +4,7 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.addon.javaee.jpa.ui;
+package org.jboss.forge.addon.javaee.jpa;
 
 /**
  * Describes what kind of ID an Entity uses.
@@ -12,6 +12,10 @@ package org.jboss.forge.addon.javaee.jpa.ui;
  * @author <a href="mailto:ch.schulz@joinout.de">Christoph "criztovyl" Schulz</a>
  */
 public enum EntityIdType {
+    /**
+     * Uses a {@link Long}.
+     */
+    LONG_PROPERTY,
     /**
      * Use an {@link javax.persistence.IdClass}.
      */
