+++ javaee/impl/src/main/java/org/jboss/forge/addon/javaee/jpa/ui/JPANewEntityCommandImpl.java
@@ -14,8 +14,8 @@ import javax.persistence.GenerationType;
 import org.jboss.forge.addon.configuration.Configuration;
 import org.jboss.forge.addon.configuration.facets.ConfigurationFacet;
 import org.jboss.forge.addon.facets.constraints.FacetConstraint;
+import org.jboss.forge.addon.javaee.jpa.EntityIdType;
 import org.jboss.forge.addon.javaee.jpa.PersistenceOperations;
-import org.jboss.forge.addon.javaee.jpa.ui.EntityIdType;
 import org.jboss.forge.addon.parser.java.facets.JavaSourceFacet;
 import org.jboss.forge.addon.projects.Project;
 import org.jboss.forge.addon.resource.DirectoryResource;
